feat/be: added owner checks on post/comment restore

+++ be/src/controllers/archive/restore-comment-from-archive.ts
@@ -2,12 +2,12 @@ import { HttpError } from 'src/api/errors/HttpError';
 import { IArchiveRepo } from 'src/types/repos/IArchiveRepo';
 import { ICommentRepo } from 'src/types/repos/ICommentRepo';
 import { createCommentHelper } from 'src/controllers/common/comment/create-comment-helper';
-import { IUserProfileRepo } from 'src/types/repos/IUserProfileRepo';
+import { EErrorCodes } from 'src/api/errors/EErrorCodes';
+
 export async function restoreCommentFromArchive(params: {
   archiveId: string;
   commentRepo: ICommentRepo;
   archiveRepo: IArchiveRepo;
-  userProfileRepo: IUserProfileRepo;
 }) {
   const commentArchive = await params.archiveRepo.getArchiveById(params.archiveId);
   
@@ -18,16 +18,7 @@ export async function restoreCommentFromArchive(params: {
     });
   }
 
-  const commentOwner = await params.userProfileRepo.getUserProfileById(commentArchive.data.userId);
-
-  if (!commentOwner) {
-    throw new HttpError({
-      statusCode: 404,
-      message: 'Comment owner not found'
-    });
-  }
-  
-  await createCommentHelper({
+  const comment = await createCommentHelper({
     commentRepo: params.commentRepo,
     data: {
       text: commentArchive.data.text,
@@ -37,6 +28,14 @@ export async function restoreCommentFromArchive(params: {
     user: commentArchive.data.user,
     postId: commentArchive.data.postId
   });
+
+  if (!comment) {
+    throw new HttpError({
+      statusCode: 404,
+      message: 'Comment owner not found',
+      errorCode: EErrorCodes.COMMENT_OWNER_NOT_FOUND
+    });
+  }
 
   await params.archiveRepo.deleteArchiveById(commentArchive.id);
